Initial Commit: - TSConfig installed as a dependency - TSConfig.json file added with some standard config - Temporarily appended some folder names with 'U' so that in the next  commit I can update the casing i.e.:  header/Header > headerU/Header > Header/Header  - Re: the above, could adjust casing in git config but would rather  leave as default
