--- src/pages/about/About.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function About(props) {
-    return (
-        <div>
-            ABOUT            
-        </div>
-    );
-}
-
-export default About;
