+++ src/pages/aboutU/About.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function About(props) {
+    return (
+        <div>
+            ABOUT            
+        </div>
+    );
+}
+
+export default About;
