--- src/pages/contact/Contact.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Contact() {
-    return (
-        <div>
-            CONTACT
-        </div>
-    );
-}
-
-export default Contact;
