+++ src/pages/contactU/Contact.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Contact() {
+    return (
+        <div>
+            CONTACT
+        </div>
+    );
+}
+
+export default Contact;
