--- src/pages/dashboard/Dashboard.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import DashboardNav from '../../components/DashboardNav/DashboardNav';
-
-function Dashboard() {
-    return (
-        <div>
-            
-            <DashboardNav />
-        </div>
-    );
-}
-
-export default Dashboard;
