+++ src/pages/dashboardU/Dashboard.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import DashboardNav from '../../components/DashboardNav/DashboardNav';
+
+function Dashboard() {
+    return (
+        <div>
+            
+            <DashboardNav />
+        </div>
+    );
+}
+
+export default Dashboard;
