--- src/pages/home/Home.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Home() {
-    return (
-        <div>
-            HOME
-        </div>
-    );
-}
-
-export default Home;
