+++ src/pages/homeU/Home.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Home() {
+    return (
+        <div>
+            HOME
+        </div>
+    );
+}
+
+export default Home;
